+++ src/key_constants.h
@@ -35,10 +35,17 @@ public:
         MAX_BECH32_TYPES
     };
 
+    enum Bech32mType {
+        TEX_ADDRESS,
+
+        MAX_BECH32M_TYPES
+    };
+
     virtual std::string NetworkIDString() const =0;
     virtual uint32_t BIP44CoinType() const =0;
     virtual const std::vector<unsigned char>& Base58Prefix(Base58Type type) const =0;
     virtual const std::string& Bech32HRP(Bech32Type type) const =0;
+    virtual const std::string& Bech32mHRP(Bech32mType type) const =0;
 };
 
 class CBaseKeyConstants : public KeyConstants {
@@ -47,11 +54,13 @@ public:
     uint32_t bip44CoinType;
     std::vector<unsigned char> base58Prefixes[KeyConstants::MAX_BASE58_TYPES];
     std::string bech32HRPs[KeyConstants::MAX_BECH32_TYPES];
+    std::string bech32mHRPs[KeyConstants::MAX_BECH32M_TYPES];
 
     std::string NetworkIDString() const { return strNetworkID; }
     uint32_t BIP44CoinType() const { return bip44CoinType; }
     const std::vector<unsigned char>& Base58Prefix(Base58Type type) const { return base58Prefixes[type]; }
     const std::string& Bech32HRP(Bech32Type type) const { return bech32HRPs[type]; }
+    const std::string& Bech32mHRP(Bech32mType type) const { return bech32mHRPs[type]; }
 };
 
 #endif // ZCASH_KEY_CONSTANTS_H